backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
email-validator==2.2.0
bcrypt==3.2.2
